smarthub/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-debug-toolbar==3.2.1
django-leaflet==0.28.1
factory-boy==3.2.0
Faker==8.11.0
idna==3.2
iniconfig==1.1.1
libpcap==1.10.0b15
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.0
paho-mqtt==1.5.1
pathspec==0.9.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.0
urllib3==1.26.6
